% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_power.R
\name{bootstrap_power_sim}
\alias{bootstrap_power_sim}
\title{Single Bootstrap Simulation}
\usage{
bootstrap_power_sim(
  boot_index,
  effect_size1,
  effect_size2,
  N,
  seed_vec,
  K_trial,
  tau2_trial,
  minority_rate
)
}
\description{
Single Bootstrap Simulation
}
\keyword{internal}
