% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filter_completed_trials}
\alias{filter_completed_trials}
\title{Filter Completed Trials}
\usage{
filter_completed_trials(studies, min_duration, min_enrollment)
}
\description{
Filter Completed Trials
}
\keyword{internal}
