% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_power.R
\name{fit_models_and_get_pvalue}
\alias{fit_models_and_get_pvalue}
\title{Fit Models and Extract P-values}
\usage{
fit_models_and_get_pvalue(dat_rr)
}
\description{
Fit Models and Extract P-values
}
\keyword{internal}
