% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_power.R
\name{generate_interventions_data}
\alias{generate_interventions_data}
\title{Generate Interventions Data}
\usage{
generate_interventions_data(K_trial)
}
\description{
Generate Interventions Data
}
\keyword{internal}
