% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_power.R
\name{generate_trial_data}
\alias{generate_trial_data}
\title{Generate Trial Data for Simulation}
\usage{
generate_trial_data(N, tau2_facility, effect_size, mean_rate)
}
\description{
Generate Trial Data for Simulation
}
\keyword{internal}
