% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_facility_distribution}
\alias{plot_facility_distribution}
\title{Plot Facility Distribution}
\usage{
plot_facility_distribution(facility_data)
}
\arguments{
\item{facility_data}{Facility analysis results}
}
\value{
ggplot object
}
\description{
Visualizes distribution of facilities across trials
}
