% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_power_results}
\alias{plot_power_results}
\title{Create Power Analysis Plot}
\usage{
plot_power_results(power_results, target_power = 0.8)
}
\arguments{
\item{power_results}{Results from power_sim function}

\item{target_power}{Target power level (default: 0.8)}
}
\value{
ggplot object
}
\description{
Generates visualization of power analysis results
}
