% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_power.R
\name{sample_data_for_analysis}
\alias{sample_data_for_analysis}
\title{Sample Participants and Trials}
\usage{
sample_data_for_analysis(all_records, N, interventions, seed_vec, boot_index)
}
\description{
Sample Participants and Trials
}
\keyword{internal}
