% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_power.R
\name{simulate_participants}
\alias{simulate_participants}
\title{Simulate Participant-Level Data}
\usage{
simulate_participants(interventions, tau2_trial, minority_rate = 0.2, seed)
}
\description{
Simulate Participant-Level Data
}
\keyword{internal}
