% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_trial_data}
\alias{generate_trial_data}
\title{Generate Trial Data for Simulation}
\usage{
generate_trial_data(
  N,
  tau2_trial,
  effect_size,
  effect_size_int,
  base_minority_rate
)
}
\description{
Generate Trial Data for Simulation
}
\keyword{internal}
